Authentication and Authorization using jwt to APIs

+++ controller/api/v1/post_api.js
@@ -19,13 +19,21 @@ module.exports.index= async function(req,res){
 module.exports.destroy = async function(req,res){
     try{
         let post= await Post.findById(req.params.id);
-        post.remove();
+        if(post.user == req.user.id){
+            post.remove();
 
-        await Comment.deleteMany({post: req.params.id});
+            await Comment.deleteMany({post: req.params.id});
 
-        return res.json(200,{
+            return res.json(200,{
             message: "Post and associated comments deleted succesfully"
-        });
+            });
+
+        }else{
+            return res.json(401,{
+                message: "You cannot delete this post!"
+            });
+        }
+        
         
     }catch(err){
         console.log("Error",err);
